👷 Add a workflow to prevent blocked PRs from being merged
To indicate that PR2 depends on PR1,
add `Depends on #1` to PR2's description

This GitHub workflow uses a GitHub Action to check for issues blocking a PR.
The term "issue" above is misleading as it can denote another PR
or a GitHub issue from the same repository.
iThis syntax does not allow cross repository PR dependencies.

See:

- GH Action: https://github.com/Levi-Lesches/blocking-issues
- Discussion about the need for such a feature in GitHub and
  the open-source workarounds in the meantime:
  https://github.com/orgs/community/discussions/4477#discussioncomment-1039627