Remove iAPS repository link
New users can and have followed this link intending to build Trio and built the wrong app.

To prevent user confusion, the link to the iAPS repository on the home screen (index.md) has been removed. The information regarding the origins of Trio is still present, but the link in the documentation is removed.